requirements.txt
Flask==2.3.2
yfinance==0.2.28
APScheduler==3.10.1
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
python-dotenv==1.0.0